move utils and types in one folder
