+++ src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const ORGS = ["Cyber", "AI", "Hack", "Innovate"];
